backend/requirements.txt
# Python Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0

# MySQL Dependencies
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0

# AI/ML Dependencies
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Additional utilities
requests==2.31.0
Pillow==10.1.0
